some sommit

+++ src/main/java/com/soulasphyxia/supersurvey/SurveyService.java
@@ -5,28 +5,50 @@ import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.Comparator;
 
 @Service
 @AllArgsConstructor
 public class SurveyService {
     private final SurveyRepository repository;
 
-    public ArrayList<Survey> getAllSurveys(String filter) {
-        if(filter == null) return repository.getAllSurveys();
-        return repository.getAllSurveys();
+
+    public ArrayList<Survey> getAllSurveys(String sort) {
+        ArrayList<Survey> surveys = repository.getAllSurveys();
+        switch (sort) {
+            case "name" -> surveys.sort(Comparator.comparing(Survey::getName));
+            case "id" -> surveys.sort(Comparator.comparing(Survey::getId));
+            default -> System.out.println("No option with such parameter.");
+        }
+        return surveys;
     }
 
     public void addSurvey(Survey survey) {
-        repository.save(survey);
+        if(!repository.containsSurveyWithId(survey.getId())) {
+            repository.save(survey);
+        }else {
+            throw new IllegalArgumentException("There is already survey with id: " + survey.getId());
+        }
     }
 
     public void editSurvey(Survey survey) {
-        repository.edit(survey);
+        if(repository.containsSurveyWithId(survey.getId())){
+            repository.edit(survey);
+        }else {
+            throw new IllegalArgumentException("There is no survey with id: " + survey.getId());
+        }
 
     }
 
     public void deleteSurvey(int id) {
-        repository.delete(id);
+        if(repository.containsSurveyWithId(id)) {
+            repository.delete(id);
+        } else{
+            throw new IllegalArgumentException("There is no survey with id: " + id);
+        }
+
 
     }
 }
